Fix the dependency situation
Dependencies were really messy, everybody was depending on something
that nobody was depending on, and having workspace dependencies shared
(for things like anyhow, serde, ron, etc. which are used by all the
projects) is better practice.

Also update iced with a special fork to avoid dependency conflict on
web-sys=0.3.67 and web-sys=0.3.68 (collision with egui-glow)

This commit also gets rid of useless dependencies in fht-compositor that
smithay already depended on anyway and reexported.