fht-compositor: Rework the window logic.
A window is now split into a surface and an actual window. The former is
responsible for managing the toplevel of the window, the latter is for
drawing the border and managing the surface, associating it with a
location, z-index, etc.

This also changed some of the window api, renaming `is_*` functions to
simply `*`.

This incidentally also pushed me to add support for windows that decide
to unmap theirselves (basically destroy their WlSurfaces, to "minimize"
themselves)