fht-compositor: Rework some rendering logic
Implement a state machine that is attached to every Output instance. It
tracks the current rendering state (idle, queued, waiting for vblank,
etc...) for this output. Queued outputs are then rendered on every
State::dispatch call.

In addition to this, we also track the current frame sequence of the
output to avoid sending multiple frame callbacks to surface in a single
frame.

This commit also moved sending dmabuf feedbacks and updating
the primary scanout output of surfaces outside of Fht::send_frames