fht-compositor: Fix screencast memory leak.
Previous commit made screencasting render its own render elements when
needed, this proved to cause memory usage to go up since the elements
were living as long as the renderer itself.

I found a solution to this and its using the render elements the backend
generates. We just have cursor_elements_len passed in so that we can
slice in the render_elements vector and skip the cursor elements if we
need to hide the cursor.

This also moved rendering screencasts in the backend side again,
allowing us to track screencopy times