fht-compositor: Re-implement fullscreen support
This was scraped in the `workspace-logic-v2` branch, but now its back!

The system is still the same as before: only one fullscreened
tile/workspace.
When focusing another tile, inserting a new one, or another one gets
fullscreened, the current loses that position and gets inserted back into
the regular tiles.