fht-compositor: Implement window closing animations.
This is still not perfect, but most of the way there.

We actually check for when a window is about to close in:
1. CompositorHandler::commit, if a surface render buffer gets removed
2. A pre-commit hook, watching for a pending change in a surface
  BufferAssignement
3. XdgShellHandler::toplevel_destroyed, obvious.

For the first change, I added root_surfaces, a cache of every root
surface in the compositor, a cool trick taken from niri (most of what
happens here is, so, eh.)

Yet this is NOT perfect, for some clients (notably libadwaita ones), the
surface view offsets do no forcibly match, and they might change their
size for no apparent reason, weird.
Test client for this case => app.drey.Elastic on flatpak

Another change that I decided todo is move the pre commit hook away from
the WorkspaceTile::new function. Testing whether an element is about to
close should not be the matter of the tile, but rather the compositor.