Implement window rules re-resolving
As a change to the migration to toml configuration format, window rules
are now stored inside the Window itself as a set of ResolvedWindowRules

This commit implements the re-resolving of these window rules, for example
when Window state changes, or when reloading the configuration

This commit also implements State::reload_config. Though it is not
perfect. It can load and apply the new configuration partially, which may
lead to some invalid state of the compositor.