Re-implement workspace switch animations
This new implementation is implemented on a per-workspace basis.
Each workspace stores a render offset which is used to draw all
the render elements with a given render offset.

When a workspace switch is triggered, the workspaces render offsets
get set to valid values in order to switch workspaces.

This allows the user to switch back and forth between workspaces
without having to wait for the switch animation to finish. Switching
to other workspaces will also be handled properly.