feat: `winit-backend`
This commits drops support for the X11 backend as its too much of a
hassle to use since I do not have Xwayland on my compositor, and instead
introduces a winit backend, which is able to run under wayland sessions.

This also renames udev_backend to udev-backend, to keep consistency with
other feature names.