Use proper `puffin` setup for profiling
This gets rid of the profiling crate and instead uses the puffin crate
directly. We still create our own crates macros to not add any useless
code if we disabled profile-with-puffin feature.

Tracy profiling is for what I need to measure overkill, puffin does the
job well enough for my purposes.