config: Add output position configuration
Now we re-arrange outputs whenever we add/remove one. The order in which
they are arranged is based on their name (for udev backend its just the
connector's name)

I should probably add a way to test the outputs, and maybe even get
another display in order to properly test cases? (Help is wanted)

This also fixes Fht::pointer_focus_target_under not accounting for output
position when returning the position of the focus target