Make a current decision to NOT support fractional scaling
Yes, we adversite the protocol, yes, smithay has f64 types, but no, it's
too complicated and not really worth it with the state of the environment
at the moment.

Even if compositor support is here, not many clients actually support it,
giving out blurry, mis-scaled, and ugly results.

In addition, a lot of smithay render elements do not support fractional
scaling, for example MemoryRenderBuffers, TextureElements, etc. Rewriting
these would be pretty tiring... I'd rather wait for support on their
behalf (which will be a million times better than anything I could come
up with)

Combine that with the fact that I do not need it too made me take a final
decision to drop support for it entierly in the compositor, starting with
the interfaces for rendering, using i32 for scale type.

Note however that custom render elements will still be using Scale<f64>
to comply with smithay's Element/RenderElement traits

This will make a lot of stuff for me way simpler down the road.