renderer/blur: Be smarter about when to render blur
Now we don't render blur below fully opaque windows, avoiding a useless
texture element.

We don't regenerate pixels for the optimized_blur_buffer unless there was
an update to the bottom/background layer shells **AND** there's a tile
that is currently requesting blur on the monitor.

This damage-tracking made BlurPrologueElement madness pointless since now
we mark the optimized_blur dirty on layer commit