space(decorations): Rework border rendering code
This commit addresses the rendering and logic of the borders that are
drawn around tiles. The biggest change is that now all Tiles have a
border field, which contains the border state and an associated
`FhtPixelShaderElement`.

Since border elements are now cached, we are properly doing damage
tracking (see #62), which is obviously a win for performance.

Another goodie introduced by this commit is border animations, they now
smoothly transition between states, including colors, border radii, and
thickness