space(decorations): Rework shadow rendering
Same thing that happened with borders is now happening with shadows,
each Tile is not associated with a unique `Shadow` element (which itself
is just a `FhtPixelShaderElement`). Tremendously helping the damage
tracker that was wrecked by new unknown render elements on each render
call. This said layer-shells still don't have this implemented since its
a little more complicated...

Fixes #62