feat(ipc): simplified broadcast event
The broadcast event which is used to update the subscribers is updated
for a simpler approach which requires 0 data.

This setup involves in the broadcast approaching the server as a client
to receive data. After receiving the data, it updates all the
subscribers with the latest data.

This strategy, although not as fast and clean as building data all by
its own, it is surely faster than polling and wayyyy simple to implement
and maintain.