wip(ipc): cleaner & less hackier broadcast
Implementing a cleaner and potentially less hacky way of broadcasting to
subscribers.

Previously, we the implementation involved the broadcast handler re-entering into the
compositor IPC as a client itself so that it can send the response of
`handle_ipc_client_request` (fn implementation [`State`] in ipc/mod.rs)
to all subscribers.

This strategy had its own pros and cons. The pros were that it was
simpler to maintain, implement and work (the full reason for the
implementation of this strategy in previous commits).

But the cons of this strategy was bigger than I anticipated. It crashed
the compositor (due to the blocking of main thread). Turns out, there is
no easy way to not make it block the main thread. I did want to use
libraries like tokio but it feels redundant for just 1 broadcast system
where every other part of the IPC use either callloop or futures_utils.
And another notable cons of this strategy (along with every other
strategy we used before) is that we have to call
`broadcast_event!(...)` macro everywhere we changed the data.

Now that we talked about the pros and cons of the "re-enter as client"
strategy (or wtvr you want to call), lets get into the pros and cons of
our new strategy.

This new strategy basically is "diffing". Yeah, it diffs the previously
stored data of the subscribers and diffs it against the new State. If
there are any changes, then it will update the subscribers.

This strategy already feels so efficient, doesn't it? It is faster, require
low to no maintainance (because we don't have to call broadcast
everywhere as it diffs data on its own to update subscribers) and seems
kinda easy to implement (but not as easy as the "re-enter as client"
strategy).

But I am anticipating unexpected results this time because everything I
try seems to go in vain.

I expect some performance costs and potential crashes. I think that both
of those are a possibility because we call the `diff_and_update`
function in the loop_handle of `start(loop_handle, wayland_socket_name)` in `ipc/mod.rs`. I also expect issues as I didn't test this change. Read the commit message, its wip.

So yeah... that's it... I will follow up with a new commit after I
finish the diffing engine.

Hopefully, this will be the end of it. I have already written/rewritten
like 900 lines of code so.... yeah... I hope this works.