feat: Removed unnecessary subscribe events
The subscribe feature has been shrunk down to just this:

```bash
$ fht-compositor ipc subscribe
```

Previously, we had an option to subscribe to individual events and now
it has been shrunk down just to a subscribe command which gives all the
response that you need.

Users can filter it out as they need.

How to format the json output?

The json result comes in a format like this if you subscribed:

{ "event": "workspace", "data": "all data here" }

You can get your event by accessing the event key of each response. If
it matches the event that you need, then you can access the data key
which holds the data of that event.