renderer(blur): Start reworking blur rendering
For now, I've removed `BlurElement`, and instead is swapping it for a
(hopefully better) dummy implementation with a `BlurRegion` and
`BlurRegionRenderElement`.

You insert the region inside whatever you want to blur (for us its
tiles and layer-shells), update it with damage, then get the render
element which does proper damage tracking.

For now, the implementation is a dummy one with no actual blur
operations, but the damage tracking is correct from testing.